backend/requirements.txt
asgiref==3.8.1
cassandra-driver==3.29.2
click==8.1.7
Django==5.1.3
django-cassandra-engine==1.9.0
django-environ==0.11.2
environ==1.0
geomet
PyYAML==6.0.2
scylla-driver==3.28.0
six==1.16.0
sqlparse==0.5.2
pyasyncore==1.0.4
django-cors-headers==4.6.0
django-cors-middleware==1.5.0
bcrypt==4.0.1
djangorestframework==3.15.2
djangorestframework-jwt
